pyproject.toml
[project]
name = "worster-underwood-cfb"
version = "0.1.0"
description = "An ensemble college football model system using both power ratings (Underwood) and resume ratings (Worster) to create a composite ranking."
authors = [
    {name = "Winston Underwood"}, 
    {name = "Hank Worster"}
]

readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "requests",
    "numpy",
    "scipy",
    "cfbd",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
parquet = ["pyarrow>=15"]
sheets = [
    "gspread>=6",
    "google-auth>=2",
]
dev = ["pytest", "black", "nbstripout", "jupyterlab", "ipykernel>=6.30.1"]

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "pytest",
  "black",
  "nbstripout",
  "jupyterlab",
  "jupyterlab-git",
  "ipykernel>=6.30.1",
]

[tool.hatch.build.targets.wheel]
packages = ["worster_underwood_cfb"]

[tool.hatch.build]
exclude = [
  "/notebooks/**",
  "/data/**",
  "/models/**",
  "/batch/**",
  "/tests/**",
]
